requirements.txt
# Core
Flask==3.0.3
Werkzeug==3.0.3

# Auth & security
Flask-Login==0.6.3
Flask-WTF==1.2.1

# Rate limiting
Flask-Limiter==3.8.0

# Environment variables
python-dotenv==1.0.1

# HTTP requests
requests==2.32.3

# Export
reportlab==4.2.2
